Added Parser class. Organized methods to do with parsing user commands and inputs into Parser class.
