+++ src/main/java/Storage.java
@@ -1,0 +1,2 @@
+public class Storage {
+}
